composite

+++ px-rupture/src/rupture/draw/sprite.hpp
@@ -9,10 +9,10 @@ namespace px {
 	struct sprite
 		: public image
 	{
-		double x_transpose;
-		double y_transpose;
-		double x_multiple;
-		double y_multiple;
+		float x_transpose;
+		float y_transpose;
+		float x_multiple;
+		float y_multiple;
 
 		unsigned int texture_index;
 		const char* name;
